Extract checkConditions into common helper
